pc - refactor RosterStudents backend into two controllers to make maintenance easier
The RosterStudentsController and tests were getting too big.
This commit refactors them into two separate controllers,
with the CSV one focusing only on CSV uploads.

We also factor out a few inner utility classes
that are now shared between both.
