+++ src/main/java/edu/ucsb/cs156/frontiers/enums/InsertStatus.java
@@ -1,0 +1,7 @@
+package edu.ucsb.cs156.frontiers.enums;
+
+public enum InsertStatus {
+  INSERTED,
+  UPDATED,
+  REJECTED
+}
