+++ src/main/java/edu/ucsb/cs156/frontiers/models/UpsertResponse.java
@@ -1,0 +1,10 @@
+package edu.ucsb.cs156.frontiers.models;
+
+import edu.ucsb.cs156.frontiers.entities.RosterStudent;
+import edu.ucsb.cs156.frontiers.enums.InsertStatus;
+
+public record UpsertResponse(InsertStatus insertStatus, RosterStudent rosterStudent) {
+  public InsertStatus getInsertStatus() {
+    return insertStatus;
+  }
+}
